It is not valid to elide lists at the root of the document (as that will miss a root tag). In addition, for lists with explicit parent tag, don't inherit the use annotation per se (the policy should be able to get it from the parent if desired). (Otherwise you get the same tag name nested).
It is not valid to elide lists at the root of the document (as that will
miss a root tag). In addition, for lists with explicit parent tag, don't
inherit the use annotation per se (the policy should be able to get it
from the parent if desired). (Otherwise you get the same tag name
nested).